requirements.txt
django
pylint
autopep8
Pillow
django-tailwind
django-crispy-forms
crispy-tailwind